Extract `options` as a separate crate.
[ci skip-build-wheels]
