+++ src/rust/engine/Cargo.toml
@@ -34,6 +34,7 @@ members = [
   "stdio",
   "logging",
   "nailgun",
+  "options",
   "process_execution",
   "process_executor",
   "rule_graph",
@@ -73,6 +74,7 @@ default-members = [
   "stdio",
   "logging",
   "nailgun",
+  "options",
   "process_execution",
   "process_executor",
   "rule_graph",
