+++ src/rust/engine/client/Cargo.toml
@@ -16,14 +16,12 @@ libc = "0.2"
 log = "0.4"
 nailgun = { path = "../nailgun" }
 nix = "0.20"
-peg = "0.7"
+options = { path = "../options" }
 sha2 = "0.9"
-shellexpand = "2.1"
 strum = "0.20"
 strum_macros = "0.20"
 sysinfo = "0.17.1"
 tokio = { version = "1.4", features = ["rt-multi-thread", "macros", "net", "io-std", "io-util"] }
-toml = "0.5"
 uname = "0.1"
 
 [dev-dependencies]
