+++ src/rust/engine/client/src/lib_tests.rs
@@ -1,7 +1,7 @@
 // Copyright 2021 Pants project contributors (see CONTRIBUTORS.md).
 // Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-use crate::render_choice;
+use options::render_choice;
 
 #[test]
 fn test() {
