+++ src/rust/engine/client/src/main.rs
@@ -37,10 +37,8 @@ use log::debug;
 use strum::VariantNames;
 use strum_macros::{AsRefStr, EnumString, EnumVariantNames};
 
-use client::option_id;
-use client::options::OptionParser;
 use client::pantsd;
-use client::render_choice;
+use options::{option_id, render_choice, OptionParser};
 
 // TODO(John Sirois): Maybe consolidate with PythonLogLevel in src/rust/engine/logging/src/lib.rs.
 #[derive(AsRefStr, EnumString, EnumVariantNames)]
