+++ src/rust/engine/options/src/lib.rs
@@ -1,9 +1,39 @@
 // Copyright 2021 Pants project contributors (see CONTRIBUTORS.md).
 // Licensed under the Apache License, Version 2.0 (see LICENSE).
 
+#![deny(warnings)]
+// Enable all clippy lints except for many of the pedantic ones. It's a shame this needs to be copied and pasted across crates, but there doesn't appear to be a way to include inner attributes from a common source.
+#![deny(
+  clippy::all,
+  clippy::default_trait_access,
+  clippy::expl_impl_clone_on_copy,
+  clippy::if_not_else,
+  clippy::needless_continue,
+  clippy::unseparated_literal_suffix,
+  // TODO: Falsely triggers for async/await:
+  //   see https://github.com/rust-lang/rust-clippy/issues/5360
+  // clippy::used_underscore_binding
+)]
+// It is often more clear to show that nothing is being moved.
+#![allow(clippy::match_ref_pats)]
+// Subjective style.
+#![allow(
+  clippy::len_without_is_empty,
+  clippy::redundant_field_names,
+  clippy::too_many_arguments
+)]
+// Default isn't as big a deal as people seem to think it is.
+#![allow(clippy::new_without_default, clippy::new_ret_no_self)]
+// Arc<Mutex> can be more clear than needing to grok Orderings:
+#![allow(clippy::mutex_atomic)]
+
 mod args;
 #[cfg(test)]
 mod args_tests;
+
+mod build_root;
+#[cfg(test)]
+mod build_root_tests;
 
 mod config;
 mod env;
@@ -27,9 +57,8 @@ use std::rc::Rc;
 use self::args::Args;
 use self::config::Config;
 use self::env::Env;
-pub use self::id::{OptionId, Scope};
-use crate::build_root::BuildRoot;
-use crate::option_id;
+pub use build_root::BuildRoot;
+pub use id::{OptionId, Scope};
 
 #[derive(Copy, Clone, Debug, Eq, PartialEq)]
 pub(crate) enum ListEditAction {
@@ -237,5 +266,14 @@ impl OptionParser {
       }
     }
     Ok(string_list)
+  }
+}
+
+pub fn render_choice(items: &[&str]) -> Option<String> {
+  match items {
+    [] => None,
+    [this] => Some(this.to_string()),
+    [this, that] => Some(format!("{} or {}", this, that)),
+    [these @ .., that] => Some(format!("{} or {}", these.join(", "), that)),
   }
 }
